+++ docs/source/doc/argparse_doc/argument_doc.rst
@@ -7,5 +7,7 @@ APITestka Argument Doc
   """
   argument:
   execute_file: str (choose action file to execute)
-  will read action file and execute
+    will read action file and execute
+  execute_dir: str (choose dir include action files)
+    read and execute all file in dir
   """
