+++ je_api_testka/requests_wrapper/request_method.py
@@ -6,20 +6,20 @@ import requests
 from je_api_testka.requests_wrapper.requests_http_method_wrapper import api_tester_method
 from requests.structures import CaseInsensitiveDict
 
-from je_api_testka.utils.exception.api_test_exceptions import APITesterGetDataException
-from je_api_testka.utils.exception.api_test_exceptions import APITesterExecuteException
+from je_api_testka.utils.exception.exceptions import APITesterGetDataException
+from je_api_testka.utils.exception.exceptions import APITesterExecuteException
 from je_api_testka.utils.get_data_strcture.get_api_data import get_api_response_data
 
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_get_data_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_get_data_error_message
 
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_get_error_message
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_put_error_message
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_delete_error_message
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_post_error_message
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_head_error_message
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_options_error_message
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_patch_error_message
-from je_api_testka.utils.exception.api_test_eceptions_tag import api_test_session_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_get_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_put_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_delete_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_post_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_head_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_options_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_patch_error_message
+from je_api_testka.utils.exception.exceptions_tag import api_test_session_error_message
 
 from je_api_testka.utils.test_record.record_test_result_class import test_record
