+++ je_api_testka/utils/assert_result/result_check.py
@@ -1,4 +1,4 @@
-from je_api_testka.utils.exception.api_test_exceptions import APIAssertException
+from je_api_testka.utils.exception.exceptions import APIAssertException
 
 
 def check_result(result_dict: dict, result_check_dict: dict):
