+++ je_api_testka/utils/html_report/html_report_generate.py
@@ -1,8 +1,8 @@
 import sys
 
 from je_api_testka.utils.test_record.record_test_result_class import test_record
-from je_api_testka.utils.exception.api_test_exceptions import HTMLException
-from je_api_testka.utils.exception.api_test_eceptions_tag import html_generate_no_data_tag
+from je_api_testka.utils.exception.exceptions import HTMLException
+from je_api_testka.utils.exception.exceptions_tag import html_generate_no_data_tag
 from threading import Lock
 
 lock = Lock()
